lib: Replace std::format() with fmt::format()
It seems that this is supported in gcc-12, which is the compiler that is
used in the github actions. Therefore, because this is an easy fix and I
would prefer not to break the action again I will leave it this way for
now. In the future I will change the compiler used by the GH action.

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>